D3D12 RHI fence logic fixes: - Stop the RHI thread holiding a critical section in FD3D12CommandListData::WaitForCompletion while we're waiting on the GPU. This has a very bad effect on RHI thread performance in GPU-bound situations.
- Simplify FD3D12CommandListData::IsComplete to fix a longstanding bug in D3D12 Fence Logic. This avoids the need for holding a lock while we wait for fences to complete.

Specifically, FD3D12CommandListData::IsComplete had over-complex logic that relied on reading the contents of the ActiveGenerations queue to determine if a generation had completed. This isn't correct, since items can be dequeued before the corresponding GPU fence is hit. Now we rely only on the LastCompleteGeneration counter (similar to the existing logic in WaitForCompletion).


#ROBOMERGE-AUTHOR: ben.woodhouse
#ROBOMERGE-SOURCE: CL 20556019 via CL 20556099 via CL 20556123
#ROBOMERGE-BOT: UE5 (Release-Engine-Staging -> Main) (v954-20466795)

[CL 20559631 by ben woodhouse in ue5-main branch]